backend/requirements/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework-stubs==3.15.3
djangorestframework_simplejwt==5.5.0
idna==3.10
mypy==1.15.0
mypy-extensions==1.0.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
ruff==0.11.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
